FIX: Multiple fixes
